added hellobar fix to hellobar implementation, also added hellobar to mobile view
