+++ src/components/Shared/_hellobar.scss
@@ -1,0 +1,44 @@
+.hellobar__container {
+  background-color: #DBE9F2;
+  min-height: 44px;
+}
+
+.hellobar__text { 
+  min-height: 100%;
+  margin-bottom: 0px;
+  text-align: center;
+  padding: 10px;
+  color: #282828;
+  font-weight: 700;
+}
+
+.hellobar__close-button {
+  position: absolute;
+  right: 20px;
+  top: 8px;
+  background-repeat: no-repeat;
+  cursor: pointer; 
+}
+
+.hellobar__postman-link {
+  height: 24px;
+  font-weight: 700;
+  color: #282828;
+  background-image: linear-gradient(180deg, transparent 50%, rgba(249,157,120,.5) 0);
+  padding-right: 4px;
+  padding-bottom: 2px;
+  background-position-y: 2px;
+  background-position-x: 4px;
+  background-repeat: no-repeat;
+  text-decoration: none;
+  transition: .18s ease;
+
+  &:hover {
+    background-position: top 7px right -4px;
+    padding-right: 4px;
+    color: #282828;
+    padding-bottom: 2px;
+    text-decoration: none;
+    background-image: linear-gradient(180deg, transparent 60%, rgba(249, 157, 120, 1) 0);
+  }
+}
