+++ src/components/_all.scss
@@ -14,6 +14,7 @@
 @import '../components/Shared/_blog-header.scss';
 @import '../components/Shared/_wordpress_wysiwyg.scss';
 @import '../components/Shared/_newsletter-form.scss';
+@import '../components/Shared/_hellobar.scss';
 
 
 @import '../components/Search/_search.scss';
